Correção produto cartesiano na pesquisa de produto edicao item 31 planilha S4 nao previstos.

+++ nds-client/src/main/java/br/com/abril/nds/controllers/cadastro/ProdutoEdicaoController.java
@@ -195,7 +195,7 @@ public class ProdutoEdicaoController extends BaseController {
 		}	
 	
 		// Pesquisar:
-		Long qtd = produtoEdicaoService.countPesquisarEdicoes(
+		Integer qtd = produtoEdicaoService.countPesquisarEdicoes(
 				filtro.getCodigo(), filtro.getNome(), intervaloLancamento, 
 				intervaloPreco, situacaoLancamento, codigoDeBarras, brinde);
